Fix test execution: Move PartialOrd tests into lib.rs
Addressed review feedback from Judge on PR #32.

Changes:
- Moved 17 PartialOrd tests from standalone tests.rs into lib.rs tests module
- Deleted tests.rs file (wasn't being compiled)
- All 45 tests now run and pass (27 original + 18 PartialOrd tests)
- Code formatted with cargo fmt

The standalone tests.rs file existed but wasn't declared as a module,
so the PartialOrd tests weren't being compiled or executed. Moving them
into the inline tests module in lib.rs fixes this issue.

Test verification:
- cargo test --package types: 45 tests passed ✓
- cargo check --package executor: Success ✓
- cargo fmt: Code formatted ✓
